crawl/requirements.txt
pandas>=2.0.0
google-generativeai>=0.3.0
openpyxl>=3.1.0
tqdm>=4.65.0 